requirements.txt
acme~=1.14.0
certbot~=1.14.0
python-transip~=0.5.0
setuptools>=1.0
zope.interface~=5.4.0
